Made con use 9-key layout on startup and removed support for persisting P1/P2 setting
The reason for removing the P1/P2 persistence is that a QMK limitation stops the 9-key P2 layer from working as the default layer. In the keymap that layer is layer index 9, and QMK uses a one-byte bitmap for storing the default layer so layer index 9 can't be stored as the default layer. To avoid doing something confusing (and to avoid writing to the EEPROM unnecessarily) we just remove all the persistence code.